Title: 자동차 단일 거래 이익 최대화하기, Time: 7ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240110/자동차 단일 거래 이익 최대화하기/max-profit-of-single-car.cpp
@@ -11,7 +11,7 @@ int main(){
         cin >> A[i];
     }
     for (int i=2;i<=n;i++){
-        int m=1000;
+        long long m=1000;
         for (int j=1;j<=i;j++){
             if (A[j]<m){
                 m=A[j];
